CThemes: move save and load of colors to CThemes class
Control of colors happens now  by using theme name (directory name)
as own setting. The actual assigning  of colors happens in CThemes class.
This avoids multiple code blocks with same purpose as previously
happened in CNeutrino and CThemes.
neutrino.conf is not longer used for color settings.
Responsible for this are color.conf files within the directories of themes.